+++ dto/backchannel_authentication_complete_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019-2020 Authlete, Inc.
+// Copyright (C) 2019-2021 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -71,4 +71,10 @@ type BackchannelAuthenticationCompleteResponse struct {
 
 	//
 	Resources []string `json:"resources,omitempty"`
+
+	//
+	ServiceAttributes []Pair `json:"serviceAttributes,omitempty"`
+
+	//
+	ClientAttributes []Pair `json:"clientAttributes,omitempty"`
 }