+++ dto/device_authorization_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019-2020 Authlete, Inc.
+// Copyright (C) 2019-2021 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -46,6 +46,9 @@ type DeviceAuthorizationResponse struct {
 	Scopes []Scope `json:"scopes,omitempty"`
 
 	//
+	DynamicScopes []DynamicScope `json:"dynamicScopes,omitempty"`
+
+	//
 	ClaimNames []string `json:"claimNames,omitempty"`
 
 	//
@@ -71,6 +74,12 @@ type DeviceAuthorizationResponse struct {
 
 	//
 	Resources []string `json:"resources,omitempty"`
+
+	//
+	ServiceAttributes []Pair `json:"serviceAttributes,omitempty"`
+
+	//
+	ClientAttributes []Pair `json:"clientAttributes,omitempty"`
 
 	//
 	Warnings []string `json:"warnings,omitempty"`